fix(alert_actor): allow alert_actor to switch both ways (blue<=>green)
Restructures alert_actor to an installable package, so that it can be more easily called and installed. Adds "path to kubernetes service file" as an adjustable parameter.
